[Refactor] Classification 5/n (#1159)
* base structure

* partly working accuracy

* init files

* working accuracy

* exact match

* mix mistake

* exact matching

* init files

* fix doctest

* mypy

* fix docs

* fix integer division

* try fixing test dependency

* fix f-string

* fix integer division

* fix integer division

* fix?

* try again

* fix pep8

* docs

* try something

+++ src/torchmetrics/metric.py
@@ -337,7 +337,7 @@ class Metric(Module, ABC):
             if reduce_fn == dim_zero_sum:
                 reduced = global_state + local_state
             elif reduce_fn == dim_zero_mean:
-                reduced = ((self._update_count - 1) * global_state + local_state) / self._update_count
+                reduced = ((self._update_count - 1) * global_state + local_state).float() / self._update_count
             elif reduce_fn == dim_zero_max:
                 reduced = torch.max(global_state, local_state)
             elif reduce_fn == dim_zero_min:
